server/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
tensorflow
pillow
python-multipart
numpy